Fix `testPresentTimeout` subtest of `TestConn`, fix possible memory leak
Regarding memory leak. In the `Write` method of the route group we call
`writePacketAsync`. It returns `errCh` to us to wait the error from, this
channel is not buffered, so `writePacketAsync`'s goroutine actually blocks
while trying to pass error to this channel. But `Write` blocks on `select`
waiting either error from `errCh` or error from write deadline struct. And
if it gets deadline error, it just returns without waiting for any error from
`errCh`, making `writePacketAsync` block forever on sending to the channel.
And we might have had a lot of such running goroutines


Former-commit-id: 5125561aaafbf13191386463a04644322e74e094