Remove conSummarry check
The commit contains changes to the logic of AppStates(). We are removing the connection summary check as whenever a vpn-client gets an error it closes all the connections to retry again. Because of this even when the vpn-client is still running AppStates() gives the app status for vpn-client as stopped. Insted depending on ProcByName() is sufficient.


Former-commit-id: f80ab5dbcb1dcb41ae6c3fe6a2053ee6b9cb5aa1