Add dmsghttp module
This commit adds dmsghttp module moving the older repetative code to one func. It also incudes the func getHTTPClient which further streamlines the code.

The service modules address-resolver, discovery, transport, dmsg, router and uptime_tracker are are reliant on the new module.

The dClient (dmsg direct client) and dmsgHTTP (dmsghttp client) are added to the Visor struct.


Former-commit-id: ff8462e67ca9f9b887b8202809f2242466577e24