Fix module initialization
This commit fixes the module initialization during the visor startup.

The Transport module is initialized before the main module as a go routine since InitConcurrent() function blocks until all dependencis are initialized so as to not wait for the stun client to initialize.

Also the v.wgStunClient.Add(1) is noved from initStunClient() to NewVisor() to mitigate any issues.