Fix isNonceValid
This commit fixes the func isNonceValid() in httpauth. The func isNonceValid() used to replace the original body of the response with the body from ioutil.NopCloser() which caused the res.Body.Close() to not work and so the stream in dmsghttp did not close.

Now instead of replacing the body of the response we instead create a new response var auxResp and send it back. Since it is a copy of the original response and not a pointer, the changes to the body of auxResp does not affect the original response.


Former-commit-id: 3cea88329b819f538ef3a54b294531afebe53361