Fix delete inactive client log
This client fixes warn log in Delete inactive clients.

We only print the warn log if the error is ont equal to connection is shut down. The Broadcast() is called from dmsgC being closed during the shutdown so it runs before the Close() of the eb which causes the warning to be printed.

Since the func Broadcast() is called before Close() and the vpn-client is already inactive by then the Delete inactive clients logic in Broadcast() closes it and the Clsoe() runs after the func Broadcast() is completed because of the mutex.