Fix Parse
This commit fixes the Parse file by removing the funcs parseV1, ensureAppDisc and the strut ParseOptions because they are not used anymore and we do not write the config to file on startup.

Parse_test.go is deleted because it is not needed since we do not write to the config with Parse.

Parse now only reads from the io reader provided to ght the conf in V1 and compares the version of the config to the version of the visor and sends back the result in the new variable compat.