Fix config gen
This commit fixes config gen. Instead of saving to and using a global variable skyenv.ConfigName we use the local variable output to save and update the path of the config as needed.

Also the global var skyenv.DMSGHTTPPath that was used to  save and update and use is chanegd by dmsgHTTPPath.

The code to get sk and hv from old config was moved under one block.

Also there was a but where we generated new pk-sk pair even if the config was read without any errors was fixed.


Former-commit-id: d0373d74ce0f912760cd6d30820bf6c162e82fa4