Fix visor config common
This commit fixes the struct common in the package visorconfig.

We reuse the field path in the common struct to flush the config instead of using it from the params of the method flush which was also removed.

We are going back to the way it was done before as it is much more cleaner to do it that way and we avoid introducing any variables.


Former-commit-id: ebbb74f2cfec80976cfeea4abccfd24f82b2eb62