Add early shutdown
This commit adds a way to cleanly shutdown the visor before all the modules are initilized.

This is done by passing the signal context with cancel to the func NewVisor and passing that context to the modules instead of a new one.

We also use the same context to check the error in the wait method for the transport module as it is the on that takes the longest to initilize can clsoe the visor with the method v.Close() inside the error statement if the context is cancelled.


Former-commit-id: 516d73801293edacda523a10ed740383562cacb6