Fix datarace
This commit fixes the datarace in the stun module by removing the waitgroup and instead using the stunReady channel along side  the method isStunReady.

The stunReady chan is used directly to block until the stun client is ready and the method isStunReady is used to check if the module is ready or not.