Update vpn client logs
This commit updates the vpn-client logs by printing the error logs with print() that uses the Stderr to write out the errors.

We also get rid of *logrus.Logger in client as it was also writing logs to Stderr instead of Stdout as we need our non error logs to be.


Former-commit-id: b0a89b040e48aa822549c1ef007fb44a5d817c84