Fix vpn-client
This commit fixes the vpn-client not working issue that was introduced by the commit 920461c since it passed eventSub to the appclient before running eventSub.OnTCPDial and eventSub.OnTCPClose which caused the vpn-client to not work as intended.

We fix it by running them first before passing the eventSub to app.NewClient.


Former-commit-id: 4d6b148c64ef663aba1a9353a96c8c6bca03b359