Add app start and stop info logs
This commit adds info logs to app start and stopped detailed state.

We add the new detailed state AppDetailedStatusStopped to every app on shutdown.

The app_state is also moved from package launcher to package appserver to get around the cyclic dependency so that we can use actual Detailed status instead of strings in Proc in the package appserver.


Former-commit-id: 06dc79b70e30feb66d7e0d514eda242091da2b2c