Send error packet on error
This commit contains the logic for sending the error packet on error in SkywireNetworker.

The method AcceptRoutes of router now sends back the route group which is used to set an error in the rg.

If the error in the routegroup is set then the error packet is sent before broadcasting close packets.


Former-commit-id: dbf7a52feac67a7c8fe596b7347cc4578a0976cf