Fix and update addRuleCmd
This commit seperates the add-rule command into three more subcommands app, fwd and intfwd. This is done in order to declutter the original implementation that only had app and intfwd where the latter was incorrectly named as fwd.

We also update printRoutingRules to fix the previously wrongly named printFwdRule and add printIntFwdRule.

We also fix the error catching in funcs parseUint, ParsePK and ParseUUID.