Feat/cli json output (#1346)
* Add global flag --json to cli

* Update pkCmd

* Add func PrintFatalError

* Minor update

* Update hvpkCmd

* Update chvpkCmd

* Update summaryCmd

* Update buildInfoCmd

* Use Catch instead of PrintFatalError

This commit removes the new function PrintFatalError and uses Catch in it's place. The func Catch is now updated to work with the new json flag.

* Add func PrintError

* Update command lsAppsCmd

* Update start, stop, setAutoStart and logs command

* Update visor execCmd

* Update lsRulesCmd

* Update rmRuleCmd

* Fix and update addRuleCmd

This commit seperates the add-rule command into three more subcommands app, fwd and intfwd. This is done in order to declutter the original implementation that only had app and intfwd where the latter was incorrectly named as fwd.

We also update printRoutingRules to fix the previously wrongly named printFwdRule and add printIntFwdRule.

We also fix the error catching in funcs parseUint, ParsePK and ParseUUID.

* Update shutdownCmd

* Update startCmd

* Update lsTypesCmd

* Update PrintTransports

* Remove depreciated flag public from tp

* Update PrintTransportEntries

* Make format

* Add changelog

* Update dmsgURLCmd

* Update visorsCmd

* Update vpnUICmd

* Update vpnURLCmd

* Update vpnListCmd

* Update vpnStartCmd and vpnStopCmd

* Update vpnStatusCmd

* Update rtfind

* Update availableServersCmd

Former-commit-id: d4e0209cdccfa3837eadbde5f89093a1223f0a03