Add new subcommand arg to visor app
Thic commit contains the new subcommand arg which is used to change various args of the apps.

This subcommand setAppAutostartCmd which was previously directly under app is now also moved under arg.

The other commands contains setAppAutostartCmd, setAppKillswitchCmd and setAppSecureCmd whcih take the app name and on/off as arguemnts.

The subcommands setAppPasscodeCmd and setAppNetworkInterfaceCmd take app name and a string as their args and have a special arg named remove that removes the arg from the said app.