Fix StartVPNClient
This commit fixes the logic in StartVPNClient where we assumed that the first app in the app list will always be vpn-client and based the logic around it. It would fail if the vpn-client wasn't the first app.

Now this has been fixed and we check the whole list of apps to see if vpn-client is available and start it only if it is.


Former-commit-id: 4e2b9963ed395f14c5bb48d1c37a76e8e510e6f5