Add build-race and update build-deploy
This commit adds the new makefile rule build-race to build binaries with -race flag to use in the docker images of e2e and integration.

We also update the rule build-deploy to use the rule host-apps-deploy instead and removing the commands to build the apps in it as they were not used.