Use gocsv
This commit uses the library gocarina/gocsv for the csv instead of encoding/csv because of it's limited functionality and the additional code needed to use it. Using gocarina/gocsv simplifies the process.

It also makes it easy to update the entries in csv where as it was a bit tedious and difficult to do the same with encoding/csv.