Update transport file logging (#1376)
* User File log store instead of mem for transports

* Update fileTransportLogStore

This commit updates the fileTransportLogStore to store the logs into a csv instead of writing json to log file.

We save the logs in just one file per day. The filename now is not the transport ip but the date.

* Test gocsv lib to read and write csv

* Update and use TpLogStore

* Use api server instead of file server

* Add endpoint for privacy json

* Fix timeout on large files

This commit fixes the timeout error on large files by removing the ReadTimeout, ReadTimeout and ReadHeaderTimeout from the http server used for serving the API over dmsg.

* Use gocsv

This commit uses the library gocarina/gocsv for the csv instead of encoding/csv because of it's limited functionality and the additional code needed to use it. Using gocarina/gocsv simplifies the process.

It also makes it easy to update the entries in csv where as it was a bit tedious and difficult to do the same with encoding/csv.

* Move today to it's own method

* Minor change

* Minor changes

* Remove todo as it is not needed

* Serve custom folder for usage of the user

* Remove test log

* Add log level to dmsgserver logs

* Minor fix

The ctx done case is moved up to prevent printing 'Cannot fetch public services: context canceled' error twice on visor shutdown.

* Fox import linting

* Read ReadHeaderTimeout to httpserver to fix lint

Former-commit-id: 6446a6e5e0fbb0cc001dfaec77c391624e3fdbad