Fix FileTransportLogStore logic
This commit fixes two things; csv update row issue when update is smaller in size than the orignal and the second issue where we used to overwrite and reset LogEntry after a restart.

Now whenever a new log entry is created with MakeLogEntry we read the Entry from the log store and if there is an entry then we add it to the new one which means the update will always be equal to or bigger than the old entry in the csv which will prevent the issue.

The second part is that now we nolonger will reset the transport data on restart and the data will now be persisted and loaded back if it's available.


Former-commit-id: 050967840394b27d3abc39e01b04942b0f44124f