Fix service discover error
After the recent changes we were sending empty info{} to the SD that is not updated which caused a json error.

Although we have set it as omitempty it stil sent empty json because VPNInfo was not a pointer.

We fix it by having VPNInfo as a pointer in Service inside pkg servicedisc and we also set the fields in VPNInfo to be omitempty.


Former-commit-id: c587965483a87846af5d5153cd1f305894ccfd99