Fix inter-day transport logging
This commit fixes the interday transport logging by resetting the LogEntry if it's a new day.

This is achieved by adding a Reset method to LogEntry first. Then we add field fileName to fileTransportLogStore check for a new day.

When a new day is detected inside the method Record insted of saving the entry to a new file of the new day we save it to the file of the old day instead. The timestamp of the entry will be of the new day indicating that the transport log was inter-day.

After we have saved this entry to the old log file we reset the LogEntry so that from the next cycle we start from zero when we are saving to the new file of today. We also set the fileName with todayFileName to indicate that the new logs can now be written to todays file.