Fix method Reset of LogEntry
This commit fixes the the method Reset by changeing the type of RecvBytes and SentBytes from uint64 to *uint64.

The issue was caused because we were resetting the copies of the fields so the reset was never regestered outside of the method.

It also contains minor fixes for everything to work with the pointers.


Former-commit-id: 86d197939601d51298cbc5dd4939261d5e04401c