Add field LogStore back to Transport in config
The field LogStore was unused after the refactore because we were only using the mem log store.

We bring back the field with Type and Location and also move the filed LogRotationInterval inside LogStore renaming it to RotationInterval.

We set the default for RotationInterval to be 7 days and we use the data from LogStore to select the Logstore Type as well as the LogStore Location.


Former-commit-id: 7d1d6f751d28320c31f4cde0390af817edce3116