Update NewClient in the package app
This commit updates the func NewClient so that it can work even if a Proc Config is not available in the env. Instead it will take the procAddr from it's parameter.

If a Proc Config will be provided in the env then the procAddr param can be left as nil.