Fix RulesMap in setup node
This commit fixes the issue where the rule used to be the same for fwd and rev when both source and destination pk is the same. This happens becase the pk is used as a key in the RulesMap and in the above case the rule is overwritten instead of creating a new one.

We fix this by changing the key to be a string of 'pk:port' instead of just the PK which makes the key unique in all cases.