Fix remote forwarding
This commit fixes the issue on the remote visor where the forwarding of the conn used to be closed abruptly when multiple requests were recieved.

We fix this by getting rid of io.Copy and instead reading a http.Request from the remote conn of the requesting visor and forwarding that request to the requested local server, and then forwarding the http.Response from the local server back to the requesting visor via the remote conn.