Fix ping data limit
This commit fixes the issue with the limited data size of read of app conn by first sending the total size of the PingMsg in the new struct PingSizeMsg.

Then in the ping module we read the size of the incoming PingMsg and read on the conn until we have received all the ping data packets then then unmarshall it.

This way the ping still remains backwards compatable without needing to change the size of noise and break backwards compatability.