Add cli command visor ping and test (#1405)
* Fix Latency in network stats

The latency in networkStats is stored as uint32 miliseconds but while converting it to time.Duration, it requires input of nanoseconds. Before we were passing in uint32 miliseconds instead of nanoseconds which caused the time.duration to be as ns and not ms.

To fix this we multiply the latencyMs with a miliseconds with a nanosecond integer count to convert our latencyMs from milisecond to nanosecond.

* Add visor ping subcommand

This commit adds the subcommand ping to the command visor. It takes one argument that is a remote PK.

This is the first version of the ping where we are pinging the skychat app on the skynet and retrieving the latency of the connection.

* Fix summarie in proc after the Latency change

The json of summary is expected to have the latency in ms but we were sending it in ns after the recent change. So we convert it back to ms while creating the summary.

* test

* Add ping app module

* Add ping method to networker interface

* Add PingRoute method to router

* Use Ping instead of Dial in API

* Fix RulesMap in setup node

This commit fixes the issue where the rule used to be the same for fwd and rev when both source and destination pk is the same. This happens becase the pk is used as a key in the RulesMap and in the above case the rule is overwritten instead of creating a new one.

We fix this by changing the key to be a string of 'pk:port' instead of just the PK which makes the key unique in all cases.

* Fix noise message authentication error

We fix this error by fixgin the noise config in PingRoute by setting the RemotePK the same as LocalPK.

* Remove test logs

* Revert test changes

* Update Ping usage in dmsg network

* Check if pk to be ping is in the route hops

* Split TestRouting into DialPing, Ping and StopPing

* Linting

* Add ping latency logic

* Add pk check

* Fix DialPing by waiting for atleast one tpm to initilize

* Fix handlePingConn

* Update ping subcommand

* Fix listening on close conn error on shutdown

* Minor initilization fix

* change ping command for get arbitrary data size

* fix rpcClient errors in new way

* add `skywire-cli visor test` command

* improve logic

* set limit on packet size in ping to 2KB

* Add PingSizeMsg

* Fix ping data limit

This commit fixes the issue with the limited data size of read of app conn by first sending the total size of the PingMsg in the new struct PingSizeMsg.

Then in the ping module we read the size of the incoming PingMsg and read on the conn until we have received all the ping data packets then then unmarshall it.

This way the ping still remains backwards compatable without needing to change the size of noise and break backwards compatability.

Co-authored-by: MohammadReza Palide <mrpalide@gmail.com>
Co-authored-by: Mohammed <79150699+mrpalide@users.noreply.github.com>