Feat/skywire app example (#1409)
* Add example server

* Update makefile

* Add example client

* Write and read packet

* Minor updates in server

* Add handleConn to example-server

* Update example client and server

* Add method Set to Addr in routing

* Use routing.Addr to read the pk:port from flags

* Move example apps to example folder

* Add examples apps to config with -m flag

* Update example app names

* Add RPC and API ReserveApp

This commit adds ReserveApp to the RPC and API which will be used to set and get a procID for an external app.

We also add Reserve in ProcManager which does Reserves a proc and proc ID and returns back the Proc ID.

* Update NewClient in the package app

This commit updates the func NewClient so that it can work even if a Proc Config is not available in the env. Instead it will take the procAddr from it's parameter.

If a Proc Config will be provided in the env then the procAddr param can be left as nil.

* Update apps to use the updated NewClient

* Rename ReserveApp to RegisterApp

* Rename Reserve to Register in ProcManager

* Change return value of RegisterApp to ProcKey

* Add procKey param to NewClient

* Add cli subcommand visor app register

* Fix method Register of procManager

This commit fixes Register by adding a go routine to wait for the conn with AwaitConn.

We also change the method awaitConn of Proc to globle and add <-p.connCh to wait for the conn to be ready.

* Update example client app to use procKey from flag

* Add DeregisterApp RPC and aPI

* Add visor app deregister cli

* Update cli readme

* Fix minor issue in visor app register

* Minor fix in visor app deregister subcommand

* Update cli readme

* Add readme for example-client-app

* Add readme for example-server-app

* Revert NewClient and use NewClientFromConfig for external apps

* Update readme of both example apps

* Add skywire app API documentation

* Update main readme

* Remove example client app from gen

* Update NewProc

This commit updates NewProc to check if conf.LogDBLoc is blank. If it is then we pass the required values as nil to accomodate the tests for Register and Deregister on ProcManager where there is no need of LogDBLoc.

* Proc manager test for Register and Deregister

* Update makefile rule install-generate to install mockery v2

* Generate ProcManager with mockery v2.14.1

* Update generate mockery command to v2.14.1

* Comment TestRegisterAndDeregisterApp

In order to write test for RPC we need to complete the todo first and move them to /pkg/app/launcher or create a mock launcher for the tests.

* Minor changes

* Add command connect

* Add api Connect

* Add method Connect to RPC

* Add visor module SkywireConnect

* Update sky_connect module to depend on the launcher module and fix initSkywireConnect.

* Minor fix

* Fix Connect API

This commit fixes the PubKey in connApp to use the remotePK instead of it's own. We also wrap the conn in appnet.WrapConn.

* Fix initSkywireConnect

* Use routing.Addr instead of appnet.Addr

* Add port forward logic to initSkywireConnect

* Add port forward logic to API Connect

* Use struct clientMsg for the initial message

* Minor change

* Minor changes

* Fix port forwarding and add srv response

* Comment closeConn

* test

* Fix local server of API Connect

This commit fixes the issue where the remoteConn used to break when multiple localConn were accepted.

To fix it we make use of a proxy server that serves a local server and redirects the requests that are made to the local server to the remote server via the remoteConn.

To create the connection between the proxy and the remote server via the remoteConn we use HTTPTransport that implements a http.RoundTripper to use skynet instead of TCP.

We use the client created with our http.RoundTripper to connect to the remote server to forward requests from proxy to remote server and transfer response from remote server to proxy.

* Remove header filtering from appnet http_transport

* Uncomment closeConn in forward

* Add server reply error logic

* Minor fix

* Minor fixes

* Fix panic and add closeChan in ServeProxy

This commit fixes 'panic: http: multiple registrations for /' caused by using http.DefaultServeMux instead of using http.NewServeMux().

It also adds a new closeChan to handleFunc so that the server and remote connection can be closed when the remote server is closed.

* Check local port availability in API Connect

* Add struct Proxy and funcs AddProxy and GetProxy

* Update Connect api and rpc to return id

* Add disconnect api and rpc

* Add ls command and rpc and api

* Add comments

* Update commands

* Change skywire connect to skywire proxy

* Fix remote forwarding

This commit fixes the issue on the remote visor where the forwarding of the conn used to be closed abruptly when multiple requests were recieved.

We fix this by getting rid of io.Copy and instead reading a http.Request from the remote conn of the requesting visor and forwarding that request to the requested local server, and then forwarding the http.Response from the local server back to the requesting visor via the remote conn.

* Fix handleFunc

This commit fixes the issue with multiple requests by adding a lock to the handleFunc as well as a sync.Once in order to close the conn only once.

* Add rpc and api for RegisterHTTPPort, DeregisterHTTPPort and ListHTTPPorts

* Add isPortRegistered check

* Add headers to ResponseWriter

* Example http server

* Fix panic

* Add api to http example

* Add cmd for register, deregister and list ports

* Rename sky proxy to sky forwarding

* Rename proxy to forwardconn

* Minor fix

* Update cli readme

* Fix remote visor conn on local visor shutdown

* Fix local visor conn on remote visor shutdown

* Add readme

* Update cli readme

* Update changelog

* change import name from skyforwarding to skyfwd

* split skyfwd to skyfwd & skyrev for local forwarding and remote reverse proxying. Switch subcommands for flags; accept but do not enforce the use of arguments

* make format

* Fix duplicate port registration

* Minor fix

* Remove port alive check from DeregisterHTTPPort

* Minor changes

* Minor change

* Return error to remote conn on error on forward

* Minor change

* Close conn instead of sending error

* shorten skyfwd / skyrev to fwd & rev

* update doc command and cmd/skywire-cli/README.md docs

* make format

Co-authored-by: Moses Narrow <36607567+0pcom@users.noreply.github.com>