Expand public autoconnect logic [WIP] (#1989)
I've significantly improved the logic for the public autoconnect.

moved autoconnect logic out of pkg/servicedisc into pkg/visor
The autoconnect is really critical logic and probably should have been in pkg/visor to begin with. This change was necessary in order to use the different api clients such as ut client, ar client & tpdisc client

maintain fresh data by limiting connection attempts to a 5 minute window.
This avoids the visor taking a long time to attempt to connect to many keys in a long list, where the data may be old

autoconnect to public visors via stcpr AND visors which are already connected to public visors via sudph
This should significantly reduce the pressure on public visors

randomize the list of visors to connect to (very important)
Avoid the situation where every visor is trying to connect to the same public visor at once!

ensure that visors are online according to the uptime tracker before attempting to connect to them
This avoids attempting to connect to visors that have subsequently gone offline

determine which transport type to use when connecting with address resolver client
This is important since the public autoconnect logic is now using multiple transport types.

moved transport type constants to it's own package (from pkg/transport/network to pkg/transport/types)
There were lots of packages importing pkg/transport/network and only using the transport type constants, so this is a significant optimization

Former-commit-id: f15f4cebd0128d1dbaf5816a6fddf7212c0291be